# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "0.1"
    },
    "paths": {
...
        "/api/incident": {
            "get": {
                "tags": [
                    "Incident"
                ],
                "summary": "Get a list of incident with dynamic filtering, sorting, pagination, and advanced search",
                "operationId": "b975e78781a2350de78f7f94e96c68c4",
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Dynamic filtering with multiple fields (e.g., `status:active|age:gt:30`)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Full-text search (e.g., `keywords`)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Sort the results by a specific field (e.g., `price` or `name`)",
                        "schema": {
                            "type": "string",
                            "example": "created_at"
                        }
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "description": "Direction of sorting (asc or desc)",
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination (default: 1)",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page (default: 10)",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/IncidentResource"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "next_page_url": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "prev_page_url": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Incident"
                ],
                "summary": "Store a new incident",
                "operationId": "659f08fa44a0fcde6391af30131e18ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IncidentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Incident created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IncidentResource"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/incident/{id}": {
            "get": {
                "tags": [
                    "Incident"
                ],
                "summary": "Get a specific incident",
                "operationId": "6a2164c7b9171a64b8289556b1a722c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Incident ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IncidentResource"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Incident"
                ],
                "summary": "Update a specific incident",
                "operationId": "34fa1beb3b1cb940df2a5844aa883cf1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Incident ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IncidentRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Incident updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IncidentResource"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Incident"
                ],
                "summary": "Delete a incident",
                "operationId": "d7527ae55638767e884ee2d193104c27",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Incident ID",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Incident deleted successfully"
                    }
                }
            }
        },
        "/api/incident/{id}/restore": {
            "put": {
                "tags": [
                    "Incident"
                ],
                "summary": "Restore a deleted incident",
                "operationId": "6a7b4f67b3710b91ebcd207623b3224a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Incident ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Incident restored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The record has been restored successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/incident/{id}/force-destroy": {
            "delete": {
                "tags": [
                    "Incident"
                ],
                "summary": "Permanently delete a incident",
                "operationId": "9650f1c4d8c4e1b5f07a4903ed47b0fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Incident ID",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Incident permanently deleted"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "IncidentRequest": {
                "required": [
                    ""
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Title Example"
                    },
                    "name": {
                        "type": "string",
                        "example": "Name Example"
                    },
                    "photo": {
                        "type": "string",
                        "example": "Photo Example"
                    },
                    "post_by_name": {
                        "type": "string",
                        "example": "Post_by_name Example"
                    },
                    "post_by_phone": {
                        "type": "string",
                        "example": "Post_by_phone Example"
                    },
                    "post_by_email": {
                        "type": "string",
                        "example": "Post_by_email Example"
                    },
                    "related_to": {
                        "type": "string",
                        "example": "Related_to Example"
                    },
                    "condition": {
                        "type": "string",
                        "example": "Condition Example"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description Example"
                    },
                    "address": {
                        "type": "string",
                        "example": "Address Example"
                    },
                    "city_id": {
                        "type": "integer",
                        "example": "16"
                    },
                    "township_id": {
                        "type": "integer",
                        "example": "17"
                    },
                    "country_id": {
                        "type": "integer",
                        "example": "95"
                    },
                    "latitude": {
                        "type": "string",
                        "example": "Latitude Example"
                    },
                    "longitude": {
                        "type": "string",
                        "example": "Longitude Example"
                    },
                    "status": {
                        "type": "string",
                        "example": "Status Example"
                    },
                    "severity": {
                        "type": "string",
                        "example": "Severity Example"
                    },
                    "other_condition": {
                        "type": "string",
                        "example": "Other_condition Example"
                    }
                },
                "type": "object"
            },
...
            "IncidentResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "48"
                    },
                    "title": {
                        "type": "string",
                        "example": "Title Example"
                    },
                    "name": {
                        "type": "string",
                        "example": "Name Example"
                    },
                    "photo": {
                        "type": "string",
                        "example": "Photo Example"
                    },
                    "post_by_name": {
                        "type": "string",
                        "example": "Post_by_name Example"
                    },
                    "post_by_phone": {
                        "type": "string",
                        "example": "Post_by_phone Example"
                    },
                    "post_by_email": {
                        "type": "string",
                        "example": "Post_by_email Example"
                    },
                    "related_to": {
                        "type": "string",
                        "example": "Related_to Example"
                    },
                    "condition": {
                        "type": "string",
                        "example": "Condition Example"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description Example"
                    },
                    "address": {
                        "type": "string",
                        "example": "Address Example"
                    },
                    "city_id": {
                        "type": "integer",
                        "example": "67"
                    },
                    "township_id": {
                        "type": "integer",
                        "example": "53"
                    },
                    "country_id": {
                        "type": "integer",
                        "example": "50"
                    },
                    "latitude": {
                        "type": "string",
                        "example": "Latitude Example"
                    },
                    "longitude": {
                        "type": "string",
                        "example": "Longitude Example"
                    },
                    "status": {
                        "type": "string",
                        "example": "Status Example"
                    },
                    "severity": {
                        "type": "string",
                        "example": "Severity Example"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-31 14:46:38"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-03-31 14:46:38"
                    },
                    "other_condition": {
                        "type": "string",
                        "example": "Other_condition Example"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token Based Authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}